adelie/pyproject.toml
[project]
name = "adelie"
version = "0.1.0"
description = "Self-Communicating Autonomous AI Loop System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "requests>=2.28.0",
]

[project.scripts]
adelie-py = "adelie.cli:main"
